feat: add validation and error handler

+++ src/main/java/com/me/example/services/TodoService.java
@@ -1,5 +1,7 @@
 package com.me.example.services;
 
+import com.me.example.dtos.TodoDto;
+import com.me.example.errors.exceptions.DataNotFoundException;
 import com.me.example.models.Todo;
 import com.me.example.repositories.TodoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,32 +17,46 @@ public class TodoService {
     private TodoRepository todoRepository;
 
     public List<Todo> getTodos() {
-        return (List<Todo>) todoRepository.findAll();
+        return todoRepository.findAll();
     }
 
     public Todo getTodoById(Long id) {
-        return todoRepository.findById(id).orElseThrow();
+        return todoRepository.findById(id).orElse(null);
     }
 
     @Transactional
-    public void createTodo(Todo todo) {
+    public void createTodo(TodoDto dto) {
+        Todo todo = Todo.builder()
+                .title(dto.getTitle())
+                .description(dto.getDescription())
+                .status(dto.getStatus())
+                .build();
+
         todoRepository.save(todo);
     }
 
     @Transactional
-    public void updateTodo(Long id, Todo data) {
-        Optional<Todo> todo = todoRepository.findById(id);
-        if (todo.isPresent()) {
-            Todo updateTodo = todo.get();
-            updateTodo.setTitle(data.getTitle());
-            updateTodo.setDescription(data.getDescription());
-            updateTodo.setStatus(data.getStatus());
-            todoRepository.save(updateTodo);
+    public void updateTodo(Long id, TodoDto data) {
+        Optional<Todo> todoOptional = todoRepository.findById(id);
+        if (todoOptional.isEmpty()) {
+            throw new DataNotFoundException("Todo not found by id " + id);
         }
+
+        Todo updateTodo = todoOptional.get();
+        updateTodo.setTitle(data.getTitle());
+        updateTodo.setDescription(data.getDescription());
+        updateTodo.setStatus(data.getStatus());
+
+        todoRepository.save(updateTodo);
     }
 
     @Transactional
     public void deleteTodoById(Long id) {
+        Optional<Todo> todoOptional = todoRepository.findById(id);
+        if (todoOptional.isEmpty()) {
+            throw new DataNotFoundException("Todo not found by id " + id);
+        }
+
         todoRepository.deleteById(id);
     }
 }
